Refactoring with the hopes of decoupling stuff
Firebase is making me :'(
