+++ app/src/main/java/xyz/kotlout/kotlout/controller/UserController.java
@@ -11,20 +11,9 @@ import java.util.regex.Pattern;
 import xyz.kotlout.kotlout.model.user.User;
 
 public class UserController {
-
   private static final String USER_COLLECTION = "users";
-  /**
-   * Pattern to validate email addresses </br> Regex from: https://emailregex.com, Accessed on
-   * Friday March 5th 2021
-   */
-  private static final Pattern EMAIL_REGEX = Pattern.compile(
-      "(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])"
-  );
-  /**
-   * Pattern to validate phone numbers </br> Notation taken from Figma diagram: 000-111-2222
-   */
-  private static final Pattern PHONE_REGEX = Pattern.compile("^[\\d\\-\\+\\(\\)]*$");
   private static final String TAG = "USER CONTROLLER";
+
   private final DocumentReference userDoc;
   private ListenerRegistration snapshotListenerRef;
   private Consumer<User> updateCallback;
@@ -38,67 +27,6 @@ public class UserController {
   public UserController(String userId) {
     userDoc = FirebaseController.getFirestore().collection(USER_COLLECTION).document(userId);
     registerSnapshotListener();
-  }
-
-  public static boolean validateEmail(String email) {
-    return email == null || email.equals("") || EMAIL_REGEX.matcher(email).matches();
-  }
-
-  public static boolean validatePhoneNumber(String phoneNumber) {
-    return phoneNumber == null || phoneNumber.equals("") || PHONE_REGEX.matcher(phoneNumber).matches();
-  }
-
-  public static void registerUser(String userUUID) {
-    if (fetchUser(userUUID) == null) {
-      User newUser = new User();
-      newUser.setUuid(userUUID);
-      FirebaseFirestore firestore = FirebaseController.getFirestore();
-      firestore.collection(USER_COLLECTION).document(newUser.getUuid()).set(newUser);
-    } else {
-      throw new RuntimeException("User already registered!");
-    }
-  }
-
-  public static User fetchUser(String userUUID) {
-    FirebaseFirestore firestore = FirebaseController.getFirestore();
-    final CollectionReference collection = firestore.collection(USER_COLLECTION);
-    final User[] readUser = {null};
-    collection.document(userUUID).get().addOnCompleteListener(snapshot -> {
-      if (snapshot.isSuccessful()) {
-        DocumentSnapshot document = snapshot.getResult();
-        if (document.exists()) {
-          Log.d("FIRESTORE", "Document snapshot: " + document.getData());
-          readUser[0] = (User) document.toObject(User.class);
-        } else {
-          Log.d("FIRESTORE", "Document snapshot not found");
-        }
-      } else {
-        Log.d("FIRESTORE", "Get failed with: " + snapshot.getException());
-      }
-    });
-    return readUser[0];
-  }
-
-  public static void setInfo(User user, String firstName, String lastName, String email,
-      String phone) {
-    user.setFirstName(firstName);
-    user.setLastName(lastName);
-    user.setEmail(email);
-    user.setPhoneNumber(phone);
-  }
-
-  public static User fetchUser(String docName, FirebaseFirestore firestore) {
-    final User[] user = {new User()};
-    firestore.collection("users").document(docName).get().addOnCompleteListener(
-        task -> {
-          if (task.isSuccessful()) {
-            DocumentSnapshot documentSnapshot = task.getResult();
-            if (documentSnapshot.exists()) {
-              user[0] = documentSnapshot.toObject(User.class);
-            }
-          }
-        });
-    return user[0];
   }
 
   /**
